+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 * [\#1302](https://github.com/cosmos/relayer/pull/1302) Avoid packet get relayed when estimated gas is higher than max gas.
 * [\#1303](https://github.com/cosmos/relayer/pull/1303) Add missing max gas amount on txf to avoid estimate less gas when simualte runTx.
 * [\#1324](https://github.com/cosmos/relayer/pull/1324) Add log-level in global config.
+* [\#1325](https://github.com/cosmos/relayer/pull/1325) Ignore only file not exist error when loadConfigFile.
 
 ## v0.9.3